backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
redis>=5.0.0
cryptography
python-dotenv
